+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Removes deprecated concepts (e.g. location and enterprise).
 - structurizr-core: Adds `Workspace.trim()` to trim a workspace of unused elements (i.e. those not associated with any views).
+- structurizr-core: Adds support for SVG image views (https://github.com/structurizr/java/issues/249).
 - structurizr-client: Removes `StructurizrClient` (use `WorkspaceApiClient` instead).
 - structurizr-import: Adds support for importing decisions managed by Log4brains.
 - structurizr-import: Adds support for importing decisions in MADR format.